feat: add dataset and window making
